% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_colormap.R
\name{as_colormap}
\alias{as_colormap}
\title{Attempts to convert an object to a ColorMap object if it is not already}
\usage{
as_colormap(pal)
}
\arguments{
\item{pal}{hex color code vector, matrix or data.frame, or ColorMap object.}
}
\value{
a \code{ColorMap} object
}
\description{
Attempts to convert an object to a ColorMap object if it is not already
}
