% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batlow_df}
\alias{batlow_df}
\title{Batlow Scientific Color Palette}
\format{
A data frame with color values:
\describe{
\item{r}{Red channel value (0-255)}
\item{g}{Green channel value (0-255)}
\item{b}{Blue channel value (0-255)}
}
}
\source{
\url{https://www.fabiocrameri.ch/colourmaps/}
}
\usage{
batlow_df
}
\description{
A perceptually uniform sequential color palette from the Scientific Color Maps
collection, created by Fabio Crameri.
}
\keyword{datasets}
