% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_interpolation.R
\name{channel_interpolation}
\alias{channel_interpolation}
\title{Fit a spline through the given values. By default it ensures the output
includes the input points}
\usage{
channel_interpolation(channel, n_out = 1000L, ensure_fit = TRUE)
}
\arguments{
\item{channel}{Numeric vector of values to interpolate.}

\item{n_out}{Integer, number of output interpolated values.}

\item{ensure_fit}{Logical, whether to ensure original points are present
in the output}
}
\value{
Numeric vector of length \code{n_out}. Spline interpolated values.
}
\description{
Fit a spline through the given values. By default it ensures the output
includes the input points
}
