% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_plot.R
\name{compare_diagnostics}
\alias{compare_diagnostics}
\title{Plot comparing diagnostics of 2 palettes}
\usage{
compare_diagnostics(pal1, pal2, pal_names = NULL)
}
\arguments{
\item{pal1}{Vector of hex colour codes, matrix or data.frame or ColorMap}

\item{pal2}{Vector of hex colour codes, matrix or data.frame or ColorMap}

\item{pal_names}{Character, name of the palettes to show}
}
\description{
Create a plots of the deltas, visual error and channels with the colors of the
palette using \code{ggplot2}
}
