% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demopal.R
\name{demopal2}
\alias{demopal2}
\title{CIEDE2000 equalized demo hex palette, making a previous interpolation
in RGB space}
\usage{
demopal2(n, m = 16, init_pal = igepal_hex, fixed_index = 2, fixed_hex = NULL)
}
\arguments{
\item{n}{number of output colors.}

\item{m}{number of colors for the intermediate interpolated palette.}

\item{init_pal}{Character vector of initial hex color values to interpolate
and equalize. Defaults to \code{igepal_hex}.}

\item{fixed_index}{Integer or NULL. Index of a color to keep unaltered.
Defaults to \code{2}.}

\item{fixed_hex}{Character or NULL. Hex color value to keep unaltered.
Must be present in \code{init_pal}. Defaults to \code{NULL}.}
}
\value{
Character length-\code{n} vector of hex color values.
}
\description{
CIEDE2000 equalized demo hex palette, making a previous interpolation
in RGB space
}
