% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_info.R
\name{df_rgb_prepare}
\alias{df_rgb_prepare}
\title{Prepare RGB data for palette_info}
\usage{
df_rgb_prepare(
  rgb_data,
  maxvalue = 1,
  channel_map = c(red = "r", green = "g", blue = "b")
)
}
\arguments{
\item{rgb_data}{Data frame or matrix with RGB values by row.}

\item{maxvalue}{Maximum value in the input RGB data (default = 1).}

\item{channel_map}{Named character vector mapping \code{"red"}, \code{"green"}, \code{"blue"}
to the corresponding column names in \code{rgb_data}.}
}
\value{
Numeric matrix of RGB values scaled to \verb{[0,1]} with columns
\code{red}, \code{green}, \code{blue}.
}
\description{
Prepare RGB data for palette_info
}
