% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plot.R
\name{diagnostic_plot}
\alias{diagnostic_plot}
\title{Plot deltas, visual error and channels}
\usage{
diagnostic_plot(pal, pal_name = NULL, show_channels = TRUE)
}
\arguments{
\item{pal}{Vector of hex colour codes, matrix or data.frame or ColorMap}

\item{pal_name}{Character, name of the palette to show}

\item{show_channels}{Logical, show channels plot or not}
}
\description{
Create a plots of the deltas, visual error and channels with the colors of the
palette using \code{ggplot2}
}
