% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_interpolation.R
\name{equalize}
\alias{equalize}
\title{Equalize n deltas after resampling a colormap}
\usage{
equalize(pal, n, anchor = NULL, n_resample = 3000L)
}
\arguments{
\item{pal}{Character vector of hex color codes, \code{ColorMap} object, \code{matrix}
or \code{data.frame}}

\item{n}{Integer, number of colors in the output \code{ColorMap}}

\item{anchor}{Hex code or index of a color to keep anchored in the output
\verb{Not yet implemented}}

\item{n_resample}{Integer, number of points of the intermediate interpolation.}
}
\value{
\code{ColorMap} object with the values interpolated in the given space
and equalized deltas.
}
\description{
Equalize n deltas after resampling a colormap
}
