% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_plot.R
\name{gradient_plot}
\alias{gradient_plot}
\title{Plot a curve with an area gradient fill}
\usage{
gradient_plot(values, pal, title, max = NULL, rescale = TRUE)
}
\arguments{
\item{values}{Numeric vector}

\item{pal}{Vector of hex color codes}

\item{title}{Name of the series}

\item{max}{Optional numeric value to fix y-axis between 0 and max}

\item{rescale}{Logical. If TRUE, rescale values to \verb{[0, 1]} but show original ticks.}
}
\description{
Create a bar plot of the deltas or visual error with the colors of the
palette using \code{ggplot2}, with optional rescaling and fixed y-axis.
}
