% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_info.R
\name{hex2palette_info}
\alias{hex2palette_info}
\title{Generate palette information from HEX colors}
\usage{
hex2palette_info(hex_vec)
}
\arguments{
\item{hex_vec}{Character vector of HEX color codes.}
}
\value{
Tibble as returned by palette_info().
}
\description{
Generate palette information from HEX colors
}
