% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{hex2rgb}
\alias{hex2rgb}
\title{Convert HEX colors to RGB matrix}
\usage{
hex2rgb(x, maxvalue = 1)
}
\arguments{
\item{x}{Character vector of HEX color codes (with or without \verb{#}),
named or unnamed.}

\item{maxvalue}{Maximum value for RGB components in the output (default = 1).}
}
\value{
Numeric matrix with columns \code{red}, \code{green}, \code{blue} and row names
as color names (or HEX codes if names are missing).
}
\description{
Convert HEX colors to RGB matrix
}
