% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{igepal_hex}
\alias{igepal_hex}
\title{IGEPAL Example Color Palette}
\format{
A character vector with 4 named HEX color codes:
\describe{
\item{PANTONE 186 C}{#C8102E}
\item{PANTONE 7461 C}{#007DBA}
\item{PANTONE Yellow}{#FEDD00}
}
}
\usage{
igepal_hex
}
\description{
A named vector of HEX color codes representing a simple categorical color palette.
}
\keyword{datasets}
