% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{igepal_hex2}
\alias{igepal_hex2}
\title{IGEPAL2 Example Color Palette}
\format{
A character vector with 4 named HEX color codes:
\describe{
\item{yellow}{#FFCC00}
\item{green}{#4EC433}
\item{soft_blue}{#0099CC}
\item{blue}{#007bc4}
\item{red}{#D81126}
}
}
\usage{
igepal_hex2
}
\description{
A named vector of HEX color codes representing a simple categorical color palette.
}
\keyword{datasets}
