% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet_colormap.R
\name{jet_colors}
\alias{jet_colors}
\title{The \code{jet} colormap}
\usage{
jet_colors(n)
}
\arguments{
\item{n}{Integer, number of output colors}
}
\value{
Character vector of hex color codes
}
\description{
The \code{jet} colormap
}
