% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_resample.R
\name{lab_lerp}
\alias{lab_lerp}
\title{Linear interpolation between two Lab points}
\usage{
lab_lerp(lab1, lab2, t)
}
\arguments{
\item{lab1}{Numeric length-3 vector (L, a, b).}

\item{lab2}{Numeric length-3 vector (L, a, b).}

\item{t}{Numeric in \link{0, 1}, interpolation parameter.}
}
\value{
Numeric length-3 vector at parameter t.
}
\description{
Linear interpolation between two Lab points
}
