% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapa_galicia}
\alias{mapa_galicia}
\title{Example Map Data}
\format{
A shapefile with associated \code{.dbf}, \code{.shx}, and \code{.prj} files.
}
\source{
Created manually for package illustration.
}
\description{
This is a shapefile included in the package under \verb{inst/extdata/map/}.
It contains a map of the municipalities of Galicia for demonstration purposes.
}
\details{
To load it:

\if{html}{\out{<div class="sourceCode">}}\preformatted{map_path <- system.file("extdata/map_gal/Concellos_IGN.shp", package = "unipals")
sf::st_read(map_path)
}\if{html}{\out{</div>}}
}
