% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_interpolation.R
\name{matrix_interpolation}
\alias{matrix_interpolation}
\title{Apply interpolation by columns to a matrix of values}
\usage{
matrix_interpolation(mat, n_out = 1000L, ensure_fit = TRUE)
}
\arguments{
\item{mat}{Numeric matrix}

\item{n_out}{Integer, number of output rows}

\item{ensure_fit}{Logical, whether to ensure original points are present
in the output}
}
\value{
Numeric matrix with \code{n_out} rows. Spline interpolated values.
}
\description{
Apply interpolation by columns to a matrix of values
}
