% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_info.R
\name{palette_info}
\alias{palette_info}
\title{Generate palette information from RGB matrix}
\usage{
palette_info(rgb_matrix)
}
\arguments{
\item{rgb_matrix}{Numeric matrix of RGB values in \verb{[0,1]}.}
}
\value{
Tibble with columns: name, hex, red, green, blue,
L, a, b, delta_2000, cum_delta_2000.
}
\description{
Generate palette information from RGB matrix
}
