% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_at_cumdeltas.R
\name{path_at_cumdeltas}
\alias{path_at_cumdeltas}
\title{Invert cumulative arc-length to Lab points at target distances
("Equalize" deltas)}
\usage{
path_at_cumdeltas(inputs, targets)
}
\arguments{
\item{targets}{Numeric vector of target cumulative lengths}

\item{resampled}{Output of map_resample().}
}
\value{
Matrix of Lab points at target positions.
}
\description{
Invert cumulative arc-length to Lab points at target distances
("Equalize" deltas)
}
