% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_channels.R
\name{plot_channels}
\alias{plot_channels}
\title{Plot channels}
\usage{
plot_channels(pal, space = c("lab", "rgb"))
}
\arguments{
\item{pal}{Vector of hex colour codes, matrix or data.frame}

\item{space}{One of "lab" or "rgb"}
}
\description{
Create a line plot of the channels in the selected space using \code{ggplot2}
}
