% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_to_pal.R
\name{rescale_to_pal}
\alias{rescale_to_pal}
\title{Rescale dataset values to reflect the perceptual error induced
by the given palette}
\usage{
rescale_to_pal(values, pal, L_direction = FALSE)
}
\arguments{
\item{values}{Numeric vector of values to rescale}

\item{pal}{ColorMap object or hex color codes vector}

\item{L_direction}{Logical, indicates if the direction of the L shifts should
be taken into account for the calculations}
}
\value{
A numeric vector of rescaled values
}
\description{
Rescale values based on perceptual differences in a color palette
}
