% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_conversion.R
\name{rgb2lab}
\alias{rgb2lab}
\title{Convert RGB matrix to CIE Lab}
\usage{
rgb2lab(rgb_matrix)
}
\arguments{
\item{rgb_matrix}{Numeric matrix of RGB values in the range \verb{[0,1]}.}
}
\value{
Numeric matrix with columns \code{L}, \code{a}, \code{b} in the CIE Lab color space.
}
\description{
Convert RGB matrix to CIE Lab
}
