% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_palettes.R
\name{scale_color_unipals_d}
\alias{scale_color_unipals_d}
\title{\code{ggplot2} compatible discrete color palette}
\usage{
scale_color_unipals_d(n, pal = demopal, ...)
}
\arguments{
\item{n}{Integer, number of colors to return.}

\item{pal}{Function, returning a vector of hex codes.}

\item{...}{}
}
\value{
\code{ggplot2} compatible discrete color palette.
}
\description{
\code{ggplot2} compatible discrete color palette
}
