% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_palettes.R
\name{scale_fill_unipals_c}
\alias{scale_fill_unipals_c}
\title{\code{ggplot2} compatible continuous fill palette}
\usage{
scale_fill_unipals_c(pal = demopal, ...)
}
\arguments{
\item{pal}{Function, returning a vector of hex codes.}

\item{...}{}
}
\value{
\code{ggplot2} compatible continuous fill gradient.
}
\description{
\code{ggplot2} compatible continuous fill palette
}
