% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_interpolation.R
\name{snap_to_fit}
\alias{snap_to_fit}
\title{Corrects the number of output points to match the input points if
spread evenly}
\usage{
snap_to_fit(n_out, n_in)
}
\arguments{
\item{n_out}{Integer, number of target points for interpolation}

\item{n_in}{Integer, number of initial points}
}
\value{
Integer, corrected \code{n_out}
}
\description{
Corrects the number of output points to match the input points if
spread evenly
}
