% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unipals-package.R
\docType{package}
\name{unipals-package}
\alias{unipals}
\alias{unipals-package}
\title{unipals: Utils For Uniform Color Palettes}
\description{
An experimental package to create and diagnose perceptually uniform color palettes.
}
\author{
\strong{Maintainer}: Jordán Soubrier \email{jordan.soubrier@gmail.com} (youruser)

}
\keyword{internal}
