% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_heatmap.R
\name{volcano_heatmap}
\alias{volcano_heatmap}
\title{Volcano heatmap}
\usage{
volcano_heatmap(pal = batlow_df)
}
\arguments{
\item{pal}{Vector of hex colour codes, matrix or data.frame}
}
\description{
Plots a heatmap of the volcano dataset using \code{ggplot2}
}
