% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano_profile.R
\name{volcano_profile}
\alias{volcano_profile}
\title{Profile comparison}
\usage{
volcano_profile(pal = jet_colors(256), signed_deltas = TRUE)
}
\arguments{
\item{pal}{Vector of hex colour codes, matrix or data.frame or ColorMap object}

\item{signed_deltas}{Logical, whether to use or not Luminance sign in the
simulation of the perceived error.}
}
\description{
Create a line plot using \code{ggplot2} that compares the real profile of
the volcano dataset against the distortion induced by a colormap.
}
