% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_resample.R
\name{delta2000_pair}
\alias{delta2000_pair}
\title{CIEDE2000 distance between two Lab rows}
\usage{
delta2000_pair(lab_i, lab_j)
}
\arguments{
\item{lab_i}{Numeric length-3 vector (L, a, b) for color i.}

\item{lab_j}{Numeric length-3 vector (L, a, b) for color j.}
}
\value{
Numeric scalar, Delta E 2000.
}
\description{
CIEDE2000 distance between two Lab rows
}
