% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoplot.R
\name{demoplot}
\alias{demoplot}
\title{A demo plot using a given palette, with the option to simulate CVDs}
\usage{
demoplot(pal = demopal, cvd = NULL)
}
\arguments{
\item{pal}{A palette function that returns hex color codes.}

\item{cvd}{Character, optional, indicates the color deficiency to emulate.
Can be 'protan', 'tritan', 'deutan', or 'desaturate'.}
}
\value{
A grid of plots demoing the palette.
}
\description{
A demo plot using a given palette, with the option to simulate CVDs
}
