% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_info.R
\name{df_rgb_prepare}
\alias{df_rgb_prepare}
\title{Prepare RGB data for palette_info}
\usage{
df_rgb_prepare(
  rgb_data,
  maxvalue = 1,
  channel_map = c(red = "red", green = "green", blue = "blue")
)
}
\arguments{
\item{rgb_data}{Data frame or matrix with RGB values by row.}

\item{maxvalue}{Maximum value in the input RGB data (default = 1).}

\item{channel_map}{Named character vector mapping "red","green","blue"
to the corresponding column names in rgb_data.}
}
\value{
Numeric matrix of RGB values scaled to \link{0,1} with columns red, green, blue.
}
\description{
Prepare RGB data for palette_info
}
