% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{igepal_hex}
\alias{igepal_hex}
\title{IGEPAL Example Color Palette}
\format{
A character vector with 4 named HEX color codes:
\describe{
\item{yellow}{#FFCC00}
\item{green}{#4EC433}
\item{blue}{#007bc4}
\item{red}{#C43E4E}
}
}
\usage{
igepal_hex
}
\description{
A named vector of HEX color codes representing a simple categorical color palette.
}
\keyword{datasets}
