% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_interpolation.R
\name{map_resample}
\alias{map_resample}
\title{Title}
\usage{
map_resample(pal, space = c("lab", "rgb"), n_out = 1000L, ensure_fit = TRUE)
}
\arguments{
\item{pal}{Character vector of hex color codes, \code{ColorMap} object, \code{matrix}
or \code{data.frame}}

\item{space}{One of "lab" or "rgb". Defaults to "lab"}

\item{n_out}{Integer, number of colors after interpolation}

\item{ensure_fit}{Logical, whether to ensure original points are present
in the output}
}
\value{
\code{ColorMap} object with the values interpolated in the given space.
}
\description{
Title
}
