% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono_gradient.R
\name{mono_gradient}
\alias{mono_gradient}
\title{Gradient palette (provisional, need to adjust L channel and correct
the function extend_palette_equal_ciede2000 for fixed extremes)}
\usage{
mono_gradient(n, from = igepal_hex[2], to = "FFFFFF")
}
\arguments{
\item{n, }{Integer, number of output colors}

\item{from}{Character, hex color code of the initial color.}

\item{to}{Character, hex color code of the end color. Defaults to 'white'.}
}
\value{
An equalized color gradient palette in hex format
}
\description{
Gradient palette (provisional, need to adjust L channel and correct
the function extend_palette_equal_ciede2000 for fixed extremes)
}
