% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano_3d.R
\name{plot_volcano_3d}
\alias{plot_volcano_3d}
\title{Plot the volcano dataset in 3D using plotly}
\usage{
plot_volcano_3d(
  pal = scico::scico(256, palette = "bamako"),
  show_contours = TRUE,
  pal_rescale = FALSE,
  z_multiplier = 1L,
  signed_deltas = FALSE,
  bw = FALSE
)
}
\arguments{
\item{pal}{Vector of hex colour codes, matrix or data.frame
Defaults to Scientific Colour Maps: "bamako"}

\item{show_contours}{Logical, whether to show contour lines (default = TRUE)}

\item{pal_rescale}{Logical, whether to rescale heights based on perceptual
differences in the palette (default = FALSE)}

\item{z_multiplier}{Numerical, multiplier to apply to the height to magnify
or decrease the perceived height.}

\item{signed_deltas}{Logical, whether to use or not Luminance sign in the
simulation of the perceived error.}

\item{bw}{Logical, indicates if the palette should be ignored and plot in
grayscale.}
}
\value{
A plotly 3D surface plot object
}
\description{
Plot the volcano dataset in 3D using plotly
}
